requirements.txt
fastapi==0.124.2
uvicorn[standard]
python-dotenv==1.2.1
facenet-pytorch==2.6.0
torch==2.2.2
torchvision==0.17.2
opencv-python==4.11.0.86
python-multipart==0.0.20
scikit-learn==1.8.0
seaborn==0.13.2
matplotlib==3.10.8
numpy==1.26.4
pillow==10.2.0
pydantic-settings
chromadb
ultralytics